pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gc-marl-iot-healing"
version = "1.0.0"
description = "OpenEnv benchmark for resilient communication infrastructure under failures and adversarial conditions"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "GC-MARL Research" }
]
license = { text = "MIT" }
dependencies = [
  "openenv-core>=0.2.0",
  "numpy>=1.24",
  "networkx>=3.0",
  "scipy>=1.10",
  "gymnasium>=0.29",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.0",
]

[tool.openenv]
name = "gc-marl-iot-healing"
manifest = "openenv.yaml"

[project.scripts]
server = "server.app:main"
